@@ -5,8 +5,9 @@
   "scripts": {
     "start": "node index.js",
     "lint": "eslint .",
-    "test": "jest"
+    "test": "node --test test/*.test.js --experimental-test-coverage"
   },
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/foxxyz/naim-monitor.git"
@@ -14,39 +15,22 @@
   "author": "Ivo Janssen <hello@ivo.la>",
   "license": "MIT",
   "devDependencies": {
-    "@appliedminds/eslint-config": "^1.5.0",
-    "eslint": "^8.43.0",
-    "eslint-plugin-jest": "^27.2.2",
-    "jest": "^29.5.0"
-  },
-  "jest": {
-    "collectCoverage": true,
-    "coverageReporters": [
-      "text"
-    ],
-    "verbose": true
+    "@appliedminds/eslint-config": "^1.6.0",
+    "eslint": "^8.56.0"
   },
   "dependencies": {
     "argparse": "^2.0.1",
     "concat-stream": "^2.0.0",
-    "fresh-console": "^1.0.4",
+    "fresh-console": "^2.0.0",
     "html-entities": "^2.4.0",
     "network-address": "^1.1.2",
     "node-ssdp": "^4.0.1",
-    "ws-plus": "^5.0.0",
-    "xml2js": "^0.6.0"
+    "ws-plus": "^6.0.2",
+    "xml2js": "^0.6.2"
   },
   "eslintConfig": {
-    "env": {
-      "es6": true,
-      "node": true,
-      "jest/globals": true
-    },
     "extends": [
       "@appliedminds"
-    ],
-    "plugins": [
-      "jest"
     ]
   }
 }
